Changing elements (e.g. move) is now reflected in both editors without saving.
This was due to the fact that every editor created it's own resource set
from the resource file (the .orm file). This caused that the editors did
not synchronized changes without saving in the respective editor (and so
triggering a reload on the resource within the other editor). This is
now obsolete as both editors work on the same resource set.